refactor: deadline description card
